MG172
Kenneth (ksgeograf) has submitted a better fix for the intersection problem. He added a call to OGR to perform real intersection in the ReadNext? call of the OGR FeatureReader?. This lets the underlying database do a simple BBOX query, and then checking the features that match the bbox for actual intersection. This can result in more features being pulled in, but effectively fixes the problem for all data sources, which the previous fix did not.