Track Id 143: Expression Functions: Fixing issues: This code drop fixes the following issues:
Aggregate Expression Function COUNT:
The return data type was set incorrectly for the function signatures that accept a string or a geometry as an input argument.

Aggregate Expression Functions MAX, MIN:
An internal flag was incorrectly set if the values were NULL values. This may result in incorrect results.

Math Expression Functions MOD, REMAINDER:
An input argument was assigned to the wrong internal variable in case the argument data type was an INT32. This resulted in incorrect values being returned by the function.