Ticket#213: Added an extended coordinate system list to the SQLServerSpatial (2008) provider.
This is a temporary solution. There is currently a discussion thread on a general solution for SRID handling by FDO providers. A more comprehensive solution might result from this discussion.

The SQL Server coordinate system catalogue (sys.spatial_reference_systems) only defines geodetic coordinate systems. However, geometry values can be associated other coordinate systems, but these are only identified by SRID. This presents a problem for the provider since it needs to be able to report the WKT for each coordinate system via FdoIGetSpatialContexts.

This submission addresses this problem by specifying a file holding the following:

    SRID number (EPSG number highly recommended)
    Coordinate System Name (Must be Mentor name for Map Create Datastore (
        non-FDO datastore) to recognize this coordinate system
    WKT (Well-Known Text)

The provider loads this file into an extended coordinate systems list (in memory). The file does not have to exist. If not present, there are no extended coordinate systems.

On reverse-engineering spatial contexts, the SRID for each the associated coordinate system for each geometric column is retrieved. If the SRID is not in the SQL Server catalogue, it is looked up in the extended coordinate system list and the WKT is retrieved.

There is one TODO item, which is to publish a file with a starting list of coordinate systems.