Ticket #390, upgrade SQLite provider to new version of 3.6.2.
This submission adds version 3.6.2 of SQLite to the vault. This is already patched with the FDO-related changes at the 5 places where that's necessary. This version of SQLite has a rewritten SQL execution engine, but the column decoding code has not changed much, since the database format hasn't changed -- so the FDO  patch carried over from 3.5.4 without much head-banging. 

I will switch the provider to use this version once I run some verification tests to make sure it still works.