Fix for ticket #389.
There were a couple of problems in the logic of Reader::Close(). The ticket describes this, but essentially, if one called Close() on a Reader which was not done reading, we did not correctly reset the state of the SQL execution engine, which was resulting in the current (and all subsequent) transaction failing to commit, which then resulted in a cascade of further problems if one keeps issuing other commands, like for example, the failure to delete features.