Ticket #390 : Updating to SQLite 3.6.2
This submission fixes a bug affecting reads of spatial query results that was caused by the SQLite upgrade. A change was needed in the way we set the next rowid in the live bytecode of the VDBE. The new engine is register based instead of stack-based so we have to set the register directly to the rowid now.