Ticket #390: Upgrading to new version of SQLite.
I am adding an experimental optimization to the SQLite B-Tree traversal code, which should significantly help the performance of our FeatureReaders. 

The stock sqlite B-Tree traversal always starts at the root when looking for the memory page containing a given feature. My observation was that in most common cases in FDO FeatureReader, features are accessed in some sort of linear or almost linear order -- for example a spatial query gives a sorted list of keys to traverse. So, it is worth checking if by any chance the feature key we are looking for exists in the page we moved to during the previous B-Tree traversal. Since we use large pages which usually fit several features, often times this is the case.

So, I am adding a check which looks at the B-Tree cursor's current location before ascending to the root and doing a full search. In my tests, the performance gain is significant, up to about 20-30% speed up for basic spatial query and read through the features.