Ticket #390: Upgrade version of SQLite.
This submission adds what I think is the last optimization to SQLite for this round. We had something similar with 3.5.4 but I had to slightly modify the optimization to reflect that the SQL engine is now register-based rather than stack based. It involves skipping a copy of data from the B-tree page to temporary memory when we are executing SQL from an FDO FeatureReader. It's a high risk optimization, but the failure mode is well understood and controlled (would result in stale pointers if multiple readers are interleaved on the same table and the data from both is used).