This submission adds some initial work on ticket #403 (implementing extended select and scrollable reader for the SQLite provider.)
The extended select and scrollable reader interfaces are the same as in SDF, but I named them with the Fdo prefix. I expect that eventually those interfaces should become standard for all providers that support scrollable readers and be exposed via a capability.

The new interfaces inherit from FdoISelect and FdoIFeatureReader, so I have folded their implementations into the standard Select and FeatureReader code -- so now all Selects with the SQLite provider will be extended and all readers scrollable (well, not quite, but it's all implemented in the same place).

I am also adding some very, very, very basic feature class inheritance support based on a patch by Philip Kast. All the code does not is flatten the schema during ApplySchema, so that inherited classes get the correct properties in their tables.