Ticket #403: Some more work on the extended select command. Implemented support for ordering of query results. The hard work is done by sqlite's ORDER BY.
I massaged the code so that the regular feature reader can be used to iterate over the ordered query results. I did it by implementing a generic iterator over row ids which can be used both to iterate over ordered results, and also for the scrollable reader. Its functioning is similar to the spatial iterator used to iterate over spatial queries, but the reader can only work with a single iterator. Because of that, in case there is both a spatial filter and ordering, some preprocessing of the result set is necessary and performed before returning the reader.