This submission changes the way the provider works with SQLite database connections, in order to allow for better read/write concurrency.
The provider will now open two sqlite connections for every FDO connection. One will be used for reads, and one for writes. This will let callers successfully commit insert transactions while also holding on to feature readers to other tables in the database. This is not possible with a single sqlite connection, because you cannot commit while holding on to an open reader (it comes down to sqlite not being able to commit from one execution engine context if there is another open execution context in the same connection, which actually makes sense from integrity viewpoint).

In addition to having two connections, we need to allow dirty reads for the reading connection, and also explicitly tell SQLite to share its memory cache between the two connections, in order to make this work. Those flags are set immediately when the read connection is opened.

So, bottom line -- you can now write to the database while holding on to an open feature reader. If you are feeling lucky, you can do that from two different threads, but that use case will require a couple of synchronization locks to really work right.