#520 Update Build. Link debug builds to debug runtime libraries.
This submission changes the Debug builds of FDO and its Providers to link against the Debug versions of the Microsoft, Apache, and ATIL runtime.

These changes primarily affect FDOCommon.

FDOCommon had previously linked to the release version of the xerces and xalan libraries to work around a structure size mismatch after upgrading to VC9. 

Afterward, we found out that the problem can also be fixed by add a preprocessor definition (_HAS_ITERATOR_DEBUGGING=0) in common.vcproj. The project that builds Xalan has the same preprocessor definition. When Transformerxalan.cpp was built without it, the structure size mismatch occurred. 

There is one additional change made by this submission. In order to be consistent, I am adding the Xalan message .hpp files to the SVN Thirdparty folder.