Fixed two bugs in the SQL statement cache, which surfaced when two statements with the same SQL were cached (due to both being used at the same time).
One bug was that all the statements were marked as "inUse" instead of just one of them, when they were being reused -- this was due to a for loop not quitting on time. 

The second bug was that the sql string was freed multiple times when clearing the cache, instead of just once, resulting in crash.