Fixed naming of spatial index backing files.
The spatial index uses temp files to store itself when it gets too big (over a million features). The file was named the same as the sqlite file and was placed in the same location. This caused two problems -- first we can't be sure that we have write access to that location, and second -- the name was not unique in case there are two feature classes with over a million features each in the same database.

To fix this, I changed the code to get a unique temp file name from Windows instead of making a name based on the original sqlite file.