Fix #596: Threading issues with nls_msg_get_W2 under heavy load
I've increased the size of the 2D array from 8 strings to 32 strings.  It's not
a perfect fix.  With a sufficiently powerful machine and under the right conditions
(32+ threads simultaneously holding strings from calls to nls_msg_get_W2) you would
have the same problem again.

I was also able to get rid of one of the two arrays.  So the overall memory usage
for the static string data went from 160kB to 320kB in Windows (twice that in Linux
since wchar's are 4 bytes there).