SHP: Turn off the nested loop check
Recently  the RELATE_NESTED_RINGS  flag has been turned ON, following some performance improvements on certain large and complex datasets.

Unfortunately it seems this is not enough. In a certain case (LandSoils.shp) the performance penalty it too high to be acceptable.

Therefore this submission is turning OFF the RELATE_NESTED_RINGS flag, pending a better solution.