Ticket#703: Submitted on behalf of Cheney Sun.
In MySQL 5.5, the value of the table_catalog column, in the information_schema tables, changed from null to "def". FdoSmPhMySQLOwner does a number of selects against information_schema tables with "table_catalog = null" in the where clause. This submission adds a check for server version. If 5.5.0 or higher, the above subclause is generated as "table_catalog = 'def'".