Ticket#739: Fixed a stack overflow which happens when reading features from XML on Linux. It occurs when a feature has a huge number of coordinates (about 97,000).
FdoXmlCoordinateGroup parses a wide-char coordinate string. On Linux, it converts each ordinate to multibyte before converting to double. Space for the intermediate multbyte was allocated on the program stack (using alloca). Items allocated by alloca are freed when the current function returns. However, the ordinates were processed in a loop, meaning that the stack could overflow if there were enough ordinates.

The fix was to add a function to convert each ordinate. This ensures that each intermediate multibyte string is free before the next one is allocated.

Some other fixes were also done to facilitate writing a unit test to cover this case:

FdoXmlFeatureReaderImpl: changed GetClassDefinition() to return the class definition for the current feature if it can be determined. This allows the FdoXmlFeatureSerializer to round trip from an FdoXmlFeatureReader to an FdoXmlFeatureWriter. 

FdoXmlFeaturePropertyReaderImpl: send the current feature's class definition via the FeatureStart event, if the class definition can be determined.

FdoXmlSchemaManager: if XML Schema mappings not set on the XML Flags that were passed in, check if they can be retrieved from the feature schemas. This provides a bit more convenience. If the app deserializes the feature schemas and passes them to an FdoXmlFeatureReader, the app no longer has to transfer the XML schema mappings from the feature schemas to the flags.