Ticket#765: Fix a bug in RFC 61 implementation for Generic RDBMS providers
This submission fixes a bug in RFC 61 implementation for SQL server spatial 
provider.

The bug is caught in save point name generating logic: it generate save points 
sp1, sp12, sp123... instead of sp1, sp2, sp3 as the code was intended to do. Now 
fix it by passing the suggested name to the string format function.