Ticket #771: Implement FDO RFC 58
This submission is the second part to implement RFC 58 (http://trac.osgeo.org/fdo/wiki/FDORfc58) - Keep It after One Long Transaction is Committed or Rollbacked. The RFC is to add an extra parameter to FdoICommitLongTransaction and 
FdoIRollbackLongTransaction commands to indicate whether to remove the long transaction after it is committed or rollbacked successfully or to keep it. 

In this submission, I extend FDO RDBMS generic framework so that it can handle newly added parameter for FdoICommitLongTransaction and FdoIRollbackLongTransaction commands.