Ticket#790: Fix the issue that a normal transaction in RDBI may be held off by an "auto-select" transaction
Submit On behalf of Sam Wang, Reviewed by Romica Dascalescu 

In the RDBI framework, transactions are ended strictly due to their nesting order. However, there are situations that between calling  tran_begin and trans_end(for non query operations), a query is executed.
 This causes the non-query transaction not the top on transaction start,  thus will fail to end. 
This submission aims at fixing this problem by skipping  transactions with ids which start with "auto-exec-" when trying to find a  matching transaction to end.