Ticket#817: GenericRdbms: modified inserts and updates to bind columns by their column type. Previously these column values were bound as strings.
Binding by string was problematic for numeric columns, when the values had fractional parts. If an insert statement sets the value for a double column to 5.55, then the provider does a locale-independent ('.' decimal point) conversion to string ("5.55") and then binds the string. The RDBMS server then converts it back to a number. However, some servers (e.g. Microsoft Access) convert using the current locale. The decimal point is lost, if the locale specifies a ',' decimal point, and the value ends up as 555.