Submit on behalf of Sam Wang
Reviewed by Romica Dascalescu<Romica.Dascalescu@autodesk.com> 
Ticket 818: WMS provider: Prevent empty CRS or SRS names added to layer's CRS name collection

This defect happens when the layer CRS name specified in the response of WMS serivce returned by "Get Capabilities" request is empty. For example,  http://wms.alaskamapped.org/bdl?version=1.3.0.

So this leads to problem, when user try to serialize the spatial context to an XML file. What we do is simply filter out empty CRS name and not add it to collection. 

Verified for both 32 bit and 64 bit provider.