StevenXu on behalf of Andy Zhang (zhanga) Integrate fix of ticket #855:  http://trac.osgeo.org/fdo/ticket/855 to Trunk.
The behavior of system call _wstat in VC9 and VC10 is different when handling file which size is bigger than 2^31. In VC9, it returns 0. But in VC10, it returns -1. Then some checks will be failed when opening a big file, which results in the big file cannot be opened in Map. 
To fix this issue, we need to use _wstat64 instead of _wstat.