#892: WFS provider doesn't work with MassGis GeoServer. Submit on behalf of Andy Zhang. Integrate revision 7109 to fdo_depot.
This submission fixed ticket #892: WFS provider doesn't work with MassGis GeoServer.
FDO WFS provider can get only 3 feature classes of the WFS server http://giswebservices.massgis.state.ma.us/geoserver/wfs. But there are about 1200 feature classes in this server. The result of command DescribeSchema is wrong. 
When executing command DescribeSchema, WFS provider sends a DescribeFeatureType request to WFS server. Then the server returns a XML which has an attribute schemaLocation. The schemaLocation is the URL for getting all the feature classes of the server. Because there are about 1200 feature classes, the server cannot handle such a large count of feature classes in a request. A 'HTTP 500 error' is returned. 
Now we split the long string to several short strings to fix the issue.

There are many feature classes which geometry types are 'MultiSurface'. It is a new type in GML 3. This submission also added support for the new type.